+++ web/angular-wallet/src/app/setup/account/account.service.ts
@@ -132,6 +132,7 @@ export class AccountService {
       const account: Account = new Account();
       // import active account
       account.type = 'active';
+      account.confirmed = false;
       const keys = generateMasterKeys(passphrase);
       const encryptedKey = encryptAES(keys.signPrivateKey, hashSHA256(pin));
       const encryptedSignKey = encryptAES(keys.agreementPrivateKey, hashSHA256(pin));
@@ -170,6 +171,7 @@ export class AccountService {
       if (existingAccount === undefined) {
         // import offline account
         account.type = 'offline';
+        account.confirmed = false;
         account.accountRS = address;
         account.account = accountId;
         await this.selectAccount(account);
@@ -273,8 +275,11 @@ export class AccountService {
       account.unconfirmedAssetBalances = remoteAccount.unconfirmedAssetBalances;
       account.balanceNQT = remoteAccount.balanceNQT;
       account.unconfirmedBalanceNQT = remoteAccount.unconfirmedBalanceNQT;
+      // @ts-ignore
+      account.confirmed = !!remoteAccount.publicKey;
     }
     catch (e) {
+      account.confirmed = false;
       console.log(e);
     }
   }