+++ lib/packages/core/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+- added `confirmed` property to `Account` model
+
 ## 0.1.1
 - added FeeQuant for `suggestFees`